# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Instantly.ai clone API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
...
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/launch:
    post:
      operationId: launchCampaign
      tags: [campaigns]
      summary: Launch / resume a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/analytics:
    get:
      operationId: getCampaignAnalytics
      tags: [campaigns]
      summary: Get analytics for a specific campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignAnalytics"
...
  /campaigns/{id}/link-clicks:
    get:
      operationId: getCampaignLinkClicks
      tags: [campaigns]
      summary: Per-link click breakdown for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of links with their click counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinkClickRow"

  # ── Sequence Steps ────────────────────────────────────────────────────────────
  /campaigns/{id}/opens:
    get:
      operationId: getCampaignOpens
      tags: [campaigns]
      summary: Opened lead breakdown for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leads that opened or otherwise engaged with the campaign email
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignOpenRow"
  /campaigns/{id}/replies:
    get:
      operationId: getCampaignReplies
      tags: [campaigns]
      summary: Replied lead breakdown for a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leads that replied to the campaign
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignReplyRow"
...
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed]
        fromName:
          type: ["string", "null"]
        replyTo:
          type: ["string", "null"]
        dailyLimit:
          type: ["integer", "null"]
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        leadsCount:
          type: integer
        sentCount:
          type: integer
        openCount:
          type: integer
        clickCount:
          type: integer
        replyCount:
          type: integer
        bounceCount:
          type: integer
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
    CampaignInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        fromName:
          type: string
        replyTo:
          type: string
        dailyLimit:
          type: integer
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
    CampaignUpdate:
      type: object
      properties:
        name:
          type: string
        fromName:
          type: string
        replyTo:
          type: string
        dailyLimit:
          type: integer
        trackOpens:
          type: boolean
        trackClicks:
          type: boolean
        includeUnsubscribe:
          type: boolean
        scheduledStartAt:
          type: ["string", "null"]
          format: date-time
    LinkClickRow:
      type: object
      required: [url, totalClicks, uniqueClicks, lastClickedAt]
      properties:
        url:
          type: string
        totalClicks:
          type: integer
        uniqueClicks:
          type: integer
        lastClickedAt:
          type: ["string", "null"]
          format: date-time
        clicks:
          type: array
          items:
            type: object
            required: [leadId, email, name, company, clickedAt]
            properties:
              leadId:
                type: integer
              email:
                type: ["string", "null"]
              name:
                type: ["string", "null"]
              company:
                type: ["string", "null"]
              clickedAt:
                type: ["string", "null"]
                format: date-time
    CampaignOpenRow:
      type: object
      required: [leadId, email, name, company, openedAt, source]
      properties:
        leadId:
          type: integer
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        openedAt:
          type: string
          format: date-time
        source:
          type: string
    CampaignReplyRow:
      type: object
      required: [leadId, email, name, company, repliedAt, subject, body]
      properties:
        leadId:
          type: integer
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        company:
          type: ["string", "null"]
        repliedAt:
          type: ["string", "null"]
          format: date-time
        subject:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
...
    CampaignAnalytics:
      type: object
      required: [campaignId, sent, opened, clicked, replied, bounced, openRate, clickRate, replyRate, bounceRate]
      properties:
        campaignId:
          type: integer
        sent:
          type: integer
        opened:
          type: integer
        clicked:
          type: integer
        uniqueClickedLeads:
          type: integer
        replied:
          type: integer
        bounced:
          type: integer
        openRate:
          type: number
        clickRate:
          type: number
        replyRate:
          type: number
        bounceRate:
          type: number

    # ── Sequence Step ─────────────────────────────────────────────────────────
...